requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
itsdangerous==2.2.0
jinja2==3.1.6
spotipy==2.23.0
starlette==0.46.2
python-dotenv==1.1.0
